dev_requirements.txt
-r requirements.txt
pillow
pre-commit
pytest
pytest-django
sphinx
tox
wheel
